chore: Moved the utils folder into the lib folder to better communicate its uses
