+++ dolphin-streak-backend/src/lib/utils/response.ts
