+++ dolphin-streak-backend/src/lib/utils/user.ts
