+++ dolphin-streak-backend/src/main.ts
@@ -1,33 +1,33 @@
-import { NestFactory } from '@nestjs/core';
-import { NestExpressApplication } from '@nestjs/platform-express';
-import { join } from 'path';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { AppModule } from './app.module';
-import { ValidationPipe } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { ApiExceptionFilter } from './utils/filters/api-exception.filter';
-import { engine } from 'express-handlebars';
+import { NestFactory } from "@nestjs/core";
+import { NestExpressApplication } from "@nestjs/platform-express";
+import { join } from "path";
+import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
+import { AppModule } from "./app.module";
+import { ValidationPipe } from "@nestjs/common";
+import { ConfigService } from "@nestjs/config";
+import { ApiExceptionFilter } from "./lib/utils/filters/api-exception.filter";
+import { engine } from "express-handlebars";
 
 async function bootstrap() {
   const app = await NestFactory.create<NestExpressApplication>(AppModule, {
     cors: {
       credentials: true,
-      origin: 'http://localhost:3000',
+      origin: "http://localhost:3000",
     },
   });
 
   const configService = app.get(ConfigService);
 
-  app.useStaticAssets(join(__dirname, '..', 'public'));
-  app.setBaseViewsDir(join(__dirname, '..', 'views'));
-  app.setViewEngine('hbs');
+  app.useStaticAssets(join(__dirname, "..", "public"));
+  app.setBaseViewsDir(join(__dirname, "..", "views"));
+  app.setViewEngine("hbs");
   app.engine(
-    'hbs',
+    "hbs",
     engine({
-      extname: 'hbs',
-      defaultLayout: 'main',
-      layoutsDir: join(__dirname, '..', 'views', 'layouts'),
-      partialsDir: join(__dirname, '..', 'views', 'partials'),
+      extname: "hbs",
+      defaultLayout: "main",
+      layoutsDir: join(__dirname, "..", "views", "layouts"),
+      partialsDir: join(__dirname, "..", "views", "partials"),
     }),
   );
 
@@ -36,14 +36,14 @@ async function bootstrap() {
   app.useGlobalFilters(new ApiExceptionFilter());
 
   const config = new DocumentBuilder()
-    .setTitle('Dolphin Streak API Endpoint')
-    .setDescription('The endpoint for Dolphin Streak API')
-    .setVersion('1.0')
+    .setTitle("Dolphin Streak API Endpoint")
+    .setDescription("The endpoint for Dolphin Streak API")
+    .setVersion("1.0")
     .addBearerAuth()
     .build();
   const documentFactory = () => SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api/docs', app, documentFactory);
+  SwaggerModule.setup("api/docs", app, documentFactory);
 
-  await app.listen(configService.get<number>('APP_PORT'));
+  await app.listen(configService.get<number>("APP_PORT"));
 }
 bootstrap();
